backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
aiosqlite>=0.20.0
sqlalchemy[asyncio]>=2.0.0
pyjwt>=2.9.0
pwdlib[argon2]>=0.2.1
python-multipart>=0.0.12
email-validator>=2.2.0
